Add the DTK_EVENT_REFRESH_LAYOUT event.
This event is used to indicate that a control needs to have the layout
of it's internal elements refresh. It should be posted in response to
anything that would require a new layout, such as a change to the size
of the control, the scaling factor, etc.

As of this commit, DTK does _not_ post this automatically.